backend/services/data_collector/requirements.txt
vnstock>=3.3.0
pandas>=2.0.0
kafka-python>=2.0.2
